pyproject.toml
[project]
name = "databricks-skeletal-streamlit-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "databricks-sdk>=0.103.0",
    "databricks-sql-connector>=4.2.5",
    "pandas>=2.3.3",
    "pyarrow>=24.0.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
